+++ projects/ng-rule-editor/src/lib/rule-editor.component.ts
@@ -1,7 +1,8 @@
-import { ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output } from '@angular/core';
+import { ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
 
 import { RuleEditorService, SimpleStyle } from './rule-editor.service';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
+import { ValidationResult } from './variable';
 
 @Component({
   // tslint:disable-next-line:component-selector
@@ -22,6 +23,8 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
   @Input() lhcStyle: SimpleStyle = {};
   @Output() save = new EventEmitter<object>();
   @Output() cancel = new EventEmitter<object>();
+
+  @ViewChild('exp') expRef;
 
   errorLoading = 'Could not detect a FHIR Questionnaire; please try a different file.';
   expressionSyntax: string;
@@ -57,6 +60,7 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
   private uneditableVariablesSubscription;
   private disableAdvancedSubscription;
   private validationSubscription;
+  private performValidationSubscription;
 
   constructor(private variableService: RuleEditorService, private liveAnnouncer: LiveAnnouncer, private changeDetectorRef: ChangeDetectorRef) {}
 
@@ -76,26 +80,52 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
     this.disableAdvancedSubscription = this.variableService.disableAdvancedChange.subscribe((disable) => {
       this.disableInterfaceToggle = disable;
     });
-    this.validationSubscription = this.variableService.validationChange.subscribe((validation) => {
-      if (validation) {
-        this.validationError = true;
-        const errorMessage = "Save button. The 'Save' button is disabled due to the validation error";
-        if (validation['section'] === "Item Variables") {
-          if (validation['variableName'])
-            this.validationErrorMessage = errorMessage + " for the variable '" + validation['name'] +
-                                          "' found in the " + validation['section'] + " section.";
-          else
-            this.validationErrorMessage = errorMessage + " found in the " + validation['section'] +
-                                          " section.";
-        } else
-          this.validationErrorMessage = errorMessage + " for the " + validation['name'] +
-                                        " found in the " + validation['section'] + " section.";
-      } else {
-        this.validationError = false;
-        this.validationErrorMessage = "";
+    this.validationSubscription = this.variableService.validationChange.subscribe((validation: ValidationResult) => {
+      this.validationError = validation.hasError;
+      this.validationErrorMessage = (this.validationError) ? this.composeAriaValidationErrorMessage(validation) : "";
+    });
+
+    // performValidationSubscription is triggered when the 'Save' button is clicked, allowing each
+    // subscribed component to validate the expression data.
+    this.performValidationSubscription = this.variableService.performValidationChange.subscribe((validation) => {     
+      // By setting the setValue to blank on simple expression that is null, empty, or undefined,
+      // it would force the validation to occurs.
+      if (this.expressionSyntax === "fhirpath" && this.finalExpression === "") {
+        this.expRef.control.markAsTouched();
+        this.expRef.control.markAsDirty();
+        this.expRef.control.setValue("");
       }
     });
   }
+
+  /**
+   * Compose the string message to be used as the aria-label to explain why the 'Save' button is disabled.
+   * @param validation - ValidationResult object which contains the validation results.
+   * @return string to be used by the 'Save' button as the aria-label in the case of any validation error.
+   */
+  composeAriaValidationErrorMessage(validation: ValidationResult): string {
+    let message = "The 'save' button is disabled due to ";
+    let itemVariablesMessage = "";
+    if (validation.errorInItemVariables) {
+      itemVariablesMessage = "error in the Item Variable section ";
+    }
+
+    let outputExpressionMessage = "";
+
+    if (validation.errorInOutputExpression) {
+      if (itemVariablesMessage !== "")
+        outputExpressionMessage += " and ";
+      outputExpressionMessage += "error with the expression in the Output Expression section";
+    } else if (validation.errorInOutputCaseStatement) {
+      if (itemVariablesMessage !== "")
+        outputExpressionMessage += " and ";
+      outputExpressionMessage += "error with the case statement in the Output Expression section";
+    }
+
+    message += itemVariablesMessage + outputExpressionMessage;
+    
+    return message;
+  };
 
   /**
    * Angular lifecycle hook called before the component is destroyed
@@ -108,6 +138,7 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
     this.disableAdvancedSubscription.unsubscribe();
 
     this.validationSubscription.unsubscribe();
+    this.performValidationSubscription.unsubscribe();
   }
 
   /**
@@ -126,6 +157,8 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
     this.variables = [];
     this.uneditableVariables = [];
     this.caseStatements = false;
+
+    this.variableService.resetValidationErrors();
 
     this.changeDetectorRef.detectChanges();
   }
@@ -180,6 +213,7 @@ export class RuleEditorComponent implements OnInit, OnChanges, OnDestroy {
    */
   export(): void {
     if (!this.validationError) {
+      this.variableService.notifyValidationCheck();
       const finalExpression = this.finalExpressionExtension;
       if (finalExpression?.valueExpression)
         finalExpression.valueExpression.expression = this.finalExpression;