+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [3.1.17] 2024-01-26
+### Added
+- Added expression validation
+- Implemented logic to disable the 'Save' button if validation fails
+
 ## [3.1.16] 2024-01-17
 ### Changed
 - Converted the Rule Editor demo to display in a dialog